Move map specific E2E tests to a subfolder
